updates event package; removes wrapped.go
