--- internal/event/wrapped.go
@@ -1,9 +1,0 @@
-package event
-
-// ReplicaSegmentWrapped wraps a ReplicationSegment with additional information
-type ReplicaSegmentWrapped struct {
-	ReplicationSegment
-	IDBatch     []string
-	SkipIDBatch []string
-	SegmentName string
-}
